+++ src/Pages/FeedPage.jsx
@@ -2,63 +2,8 @@ import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
 import MainHeader from '../Components/Common/MainHeader';
 import BottomNav from '../Components/Common/BottomNav';
-import Post from '../Components/Feed/Post';
-import axios from 'axios';
-
-
-// 피드 fetching, 게시글생성
-const MainFeed = () => {
-  const URL = 'https://api.mandarin.weniv.co.kr/';
-  const FollowingPOST = 'post/feed';
-  const Authorization =
-    'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY0NzZkNzZjYjJjYjIwNTY2MzJjZmZlYiIsImV4cCI6MTY5MDY5NDM4MCwiaWF0IjoxNjg1NTEwMzgwfQ.Bjwk8EyTTxyFP8-QYiY1SlXsAXTAYQ_Fwmi-nJ-NDx4';
-
-  const [fetchData, setfetchData] = useState(false);
-  const [FeedList, setFeedList] = useState([]);
-
-  useEffect(() => {
-    getFeed();
-  }, []);
-
-  const instance = axios.create({
-    baseURL: URL,
-    headers: {
-      'Content-Type': 'application/json',
-      Authorization: Authorization,
-    },
-  });
-
-  async function getFeed() {
-    try {
-      const response = await instance.get(FollowingPOST);
-      const FeedListArr = [];
-      for (let i = 0; i < response.data.posts.length; i++) {
-        FeedListArr.push(response.data.posts[i]);
-      }
-      console.log(FeedListArr);
-      setFeedList(FeedListArr);
-      setfetchData(true);
-    } catch (error) {
-      console.error(error);
-    }
-  }
-
-  return (
-    <>
-      <Post fetchData={fetchData} FeedList={FeedList} />
-    </>
-  );
-};
-
-// 태그버튼 컴포넌트
-const TagButton = ({ text, active, onClick }) => {
-  console.log(active);
-  return (
-    <STagButton onClick={onClick} className={active ? 'active' : ''}>
-      {text}
-    </STagButton>
-  );
-};
+import FetchFeed from '../Components/Feed/FetchFeed';
+import TagButton from '../Components/Feed/TagButton';
 
 // 피드 메인 페이지
 const FeedPage = () => {
@@ -67,8 +12,8 @@ const FeedPage = () => {
   const handleClick = index => {
     setActiveIndex(index);
   };
-
   const tagItem = ['#전체', '#스터디 모집', '#질문있어요!', '#자유게시판'];
+
   return (
     <>
       <SMain>
@@ -80,7 +25,7 @@ const FeedPage = () => {
                 <TagButton key={index} text={item} active={index === activeIndex} onClick={() => handleClick(index)} />
               ))}
             </STagLayout>
-            <MainFeed />
+            <FetchFeed />
           </>
         ) : (
           <div>검색창</div>
@@ -92,19 +37,6 @@ const FeedPage = () => {
 };
 export default FeedPage;
 
-// TagButton =====================================================
-const STagButton = styled.div`
-  border-radius: 44px;
-  border: 1px solid var(--darkgray);
-  padding: 5px 8px;
-  font-size: 12px;
-  color: var(--lightgray);
-  cursor: pointer;
-  &.active {
-    color: var(--white);
-    background-color: var(--point-color);
-  }
-`;
 // FeedPage =====================================================
 const SMain = styled.div`
   max-width: 390px;
